backend/requirements.txt
fastapi[standard]
sqlmodel
pydantic
sqlalchemy
sqlite-migrate
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
